// src/api/model/invoiceListByCompanySortBy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export type InvoiceListByCompanySortBy = typeof InvoiceListByCompanySortBy[keyof typeof InvoiceListByCompanySortBy];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceListByCompanySortBy = {
  createdDate: 'createdDate',
  duzpDate: 'duzpDate',
  dueDate: 'dueDate',
  number: 'number',
  totalWithTax: 'totalWithTax',
  status: 'status',
} as const;