// src/api/model/invoiceListByCompanyStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export type InvoiceListByCompanyStatus = typeof InvoiceListByCompanyStatus[keyof typeof InvoiceListByCompanyStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceListByCompanyStatus = {
  DRAFT: 'DRAFT',
  ISSUED: 'ISSUED',
  OVERDUE: 'OVERDUE',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED',
} as const;